Make it possible to pass HTML to the input
This is equivalent to `options` in the Rails helpers, but we kinda have
to be more specific in our naming here, since we'll likely end up having
both `input_attributes` and `label_attributes`.

As for options vs attributes, the HTML helpers use `options`, but HTML
calls them attributes.

We may want to consider using `attributes` for the HTML attributes that
should be applied to the parent/main element of the component with
`*_attributes` reserved for the child elements.