Add missing end margin
The purple button example at
https://flowbite.com/docs/components/buttons/#default-button does not
have an end margin, whereas the other buttons do, probably because it is
the last button of the group.

For consistencys sake, I've added the me-2 class to the purple button as
well, but the real question is, should the buttons come with margins at
all? The fact that me-2 was missing here seems to indicate that it is
not an innate part of the button, and the other buttons solely have the
margin classes because they need to look pretty in the documentation.