Remove margins from button
In commit c2ff3ed8e9b23d704163657bc4be20c2f91c40e9 we added a "missing"
margin to the purple style button. We did so for consistencys sake, but
the fact that the me-2 class was missing from one button seems to
indicate that the margin is not an innate part of the button, and the
other buttons solely have the margin classes because they need to look
pretty in the documentation.

The purple button example at
https://flowbite.com/docs/components/buttons/#default-button does not
have an end margin, whereas the other buttons do, probably because it is
the last button of the group - which means it is entirely coincidental
and specific to that example.

Generally speaking, margins shouldn't be a part of each component. Leave
it up to the consumer to add the whitespace around your components that
you need them to.