Add select specific options to Flowbite select input component
This adds :include_blank and :multiple options to the Flowbite select input component,
and updates the corresponding InputField component and tests accordingly.

I am omitting the :prompt option, as it is not clear it works with the select helper.
Tests seems to suggest it does not, but maybe with a select_tag helper it would.